refactor; move out auth classes
